requirements.txt
python-telegram-bot==13.8
requests==2.25.1